Adds support for Markdown and HTML message formats
Lets the user specify a format to send the telegram message in. Can be one of
`text`, `markdown`, or `html`. Also adds some bare-bones docs for the feature.

This commit bumps the minimum version of python-telegram-bot to 2.8.3, as that
was the version which introduced Markdown formatting support. That version was
released in September of 2015, and the current version of that library is 5.3,
so this should be a fairly safe minimum to enforce.